+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add IE11 support
 - `ZodError.flatten` now optionally accepts a map function for customizing the output
 - `.void()` now only accepts undefined, not null.
+- `z.enum` now supports `Readonly` string tuples
 
 ### 3.5